requirements.txt
flask
python-keycloak
Flask-SQLAlchemy
mariadb
python-dotenv
gunicorn
gevent